requirements.txt
alembic==1.12.1
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bottle==0.12.25
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
colorlog==6.7.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
fonttools==4.39.3
greenlet==3.0.1
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
joblib==1.3.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
optuna==3.4.0
optuna-dashboard==0.13.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.23
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.8.0
urllib3==1.26.15
wcwidth==0.2.6
wsproto==1.2.0
zipp==3.15.0
